Refactor VFS read to return ssize_t and implement O_NONBLOCK in TTY
- Created `include/fcntl.h` with `O_NONBLOCK` and standard flags.
- Refactored `read_type_t` and `read_fs` in `include/fs/vfs.h` and `kernel/fs/vfs.c` to return `ssize_t` to support negative error codes.
- Updated `tty_read` in `kernel/drivers/tty.c` to check for `O_NONBLOCK` and `keyboard_has_input()`, returning `-EAGAIN` if no input is available.
- Updated syscall handlers in `kernel/arch/x86_64/syscall.c` to support `O_NONBLOCK` propagation and handle signed read results.
- Updated read function signatures in all filesystem implementations: `fat32.c`, `initrd.c`, `devfs.c`, `procfs.c`, `jfs.c`, and `partition.c`.
- Updated unit tests (`test_vfs_leak.c`, `test_readv_security.c`, `test_partition.c`) to reflect API changes.
- Fixed missing `vmm_check_user_string` mock in `test_readv_security.c`.

+++ kernel/fs/initrd.c
@@ -39,7 +39,7 @@ static initrd_dir_t* find_virtual_dir(const char* name) {
     return NULL;
 }
 
-uint32_t initrd_read(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
+ssize_t initrd_read(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
     initrd_file_header_t header = file_headers[node->inode];
     if (offset > header.size)
         return 0;
